Moved shared defs to psma module
